+++ src/ui/components/transactions/FinalTransaction.tsx
@@ -23,16 +23,15 @@ export class FinalTransaction extends React.PureComponent {
     readonly props: any;
 
     render() {
-        const { transactionStatus, hasNewMessages } = this.props;
+        const { transactionStatus, hasNewMessages, message, selectedAccount } = this.props;
         const isApprove = !!transactionStatus.approveOk;
         const isReject = !!transactionStatus.rejectOk;
         const isError = !!transactionStatus.approveError;
-        const message = transactionStatus.approveOk && transactionStatus.approveOk.message || {};
         const signData = this.props.signData;
         const config = this.props.config;
         const FinalComponent = config.final;
         const isSend = message.broadcast;
-        const network = message.account && message.account.networkCode;
+        const network = selectedAccount && selectedAccount.networkCode;
         const txLink = `https://${ network === 'T' ? 'testnet.' : ''}wavesexplorer.com/tx/${message.messageHash}`;
         const className = cn(styles.txBigIcon, 'margin-main', {
             'tx-reject-icon': isReject,