+++ src/ui/components/transactions/SetAssetScript.tsx
@@ -3,34 +3,15 @@ import * as React from 'react'
 import { translate, Trans } from 'react-i18next';
 import { SignClass } from './SignClass';
 import { TxIcon } from './TransactionIcon';
-import { Balance, Button, Copy, DateFormat, Modal } from '../ui';
+import { Balance, DateFormat, Modal, ShowScript } from '../ui';
 import { TransactionBottom } from './TransactionBottom';
-import cn from 'classnames';
 import { I18N_NAME_SPACE } from '../../appConfig';
 
 @translate(I18N_NAME_SPACE)
 export class SetAssetScript extends SignClass {
     
-    scriptEl: HTMLDivElement;
-    readonly state = { showAllScript: false, showCopied: false, showResizeBtn: false };
-    _t;
-    
-    toggleShowScript = () => this.setState({ showAllScript: !this.state.showAllScript });
-    onCopy = () => this._onCopy();
-    getScriptRef = (ref) => this.scriptEl = ref;
-    
-    componentDidMount() {
-        const scrollHeight = this.scriptEl.scrollHeight;
-        const height = this.scriptEl.offsetHeight;
-        const showResizeBtn = scrollHeight > height;
-        this.setState({ showResizeBtn })
-    }
-    
     render() {
         const { data: tx } = this.props.signData;
-        const showAllClass = cn({
-            [styles.showAllScript]: this.state.showAllScript
-        });
         
         return <div className={styles.transaction}>
             {super.render()}
@@ -43,24 +24,8 @@ export class SetAssetScript extends SignClass {
                 <div className="headline2 center margin-main-big">
                     <Trans i18nKey='transactions.setAssetScriptHeader'>Set Asset Script transaction</Trans>
                 </div>
-
-                <div className={`plate plate-with-controls break-all ${showAllClass}`}>
-                    <div ref={this.getScriptRef} className={`${styles.txValue}`}>{tx.script}</div>
-                    <div className="buttons-wrapper">
-                        { tx.script ? <Copy text={tx.script} onCopy={this.onCopy}>
-                            <Button>
-                                <Trans i18nKey='transactions.copy'>Copy</Trans>
-                            </Button>
-                        </Copy> : null }
-                        { this.state.showResizeBtn ? <Button onClick={this.toggleShowScript}>
-                            {
-                                !this.state.showAllScript ?
-                                <Trans i18nKey='transactions.showAll'>Show all</Trans>:
-                                <Trans i18nKey='transactions.hide'>Hide</Trans>
-                            }
-                        </Button>: null }
-                    </div>
-                </div>
+    
+                <ShowScript script={tx.script} showNotify={true}/>
                 
                 <div className={`${styles.txRow} margin-main`}>
                     <div className="tx-title tag1 basic500">
@@ -100,11 +65,5 @@ export class SetAssetScript extends SignClass {
                 </div>
             </Modal>
         </div>
-    }
-    
-    protected _onCopy() {
-        this.setState({ showCopied: true });
-        clearTimeout(this._t);
-        this._t = setTimeout(() => this.setState({ showCopied: false }), 1000);
     }
 }