+++ changeLog.md
@@ -5,4 +5,11 @@
 + Expiration requests
 + Add black/white lists of origins
 + Add custom origins permissions
++ Add edit permissions page
++ Add smart asset transactions
++ Some design fixes
++ Refactor current active message
++ Add to keeper public state tx versions
 + Update signature lib
++ Update dictionary JSON
++ Update api massTransfer tx format